Separate impulse module from models module
