+++ src/prfmodel/models/__init__.py
@@ -2,11 +2,9 @@
 
 from .base import BaseComposite
 from .base import BaseEncoder
-from .base import BaseImpulse
-from .base import BaseModel
 from .base import BaseResponse
 from .base import BaseTemporal
-from .base import BatchDimensionError
+from .base import ModelProtocol
 from .css import Gaussian2DCSSPRFModel
 from .css import init_css_from_gaussian
 from .div_norm import DivNormGaussian2DPRFModel
@@ -25,11 +23,8 @@ from .gaussian import GaussianCFModel
 __all__ = [
     "BaseComposite",
     "BaseEncoder",
-    "BaseImpulse",
-    "BaseModel",
     "BaseResponse",
     "BaseTemporal",
-    "BatchDimensionError",
     "CFStimulusEncoder",
     "CompressiveEncoder",
     "DivNormGaussian2DPRFModel",
@@ -38,6 +33,7 @@ __all__ = [
     "Gaussian2DCSSPRFModel",
     "Gaussian2DPRFModel",
     "GaussianCFModel",
+    "ModelProtocol",
     "PRFStimulusEncoder",
     "encode_prf_response",
     "init_css_from_gaussian",
