+++ src/prfmodel/models/base.py
@@ -1,92 +1,18 @@
 """Model base classes."""
 
-from abc import ABC
 from abc import abstractmethod
-from collections.abc import Sequence
 from typing import Generic
 from typing import TypeVar
 import pandas as pd
-from keras import ops
 from prfmodel._docstring import doc
 from prfmodel.stimuli import Stimulus
 from prfmodel.typing import Tensor
-from prfmodel.utils import _get_norm_fun
+from prfmodel.utils import ModelProtocol
 
 S = TypeVar("S", bound=Stimulus)
 
 
-class BatchDimensionError(Exception):
-    """
-    Exception raised when arguments have different sizes in the batch (first) dimension.
-
-    Parameters
-    ----------
-    arg_names: Sequence[str]
-        Names of arguments that have different sizes in batch dimension.
-    arg_shapes: Sequence[tuple of int]
-        Shapes of arguments that have different sizes in batch dimension.
-
-    """
-
-    def __init__(self, arg_names: Sequence[str], arg_shapes: Sequence[tuple[int, ...]]):
-        names = ", ".join(arg_names)
-        shapes = ", ".join([str(s[0]) for s in arg_shapes])
-
-        super().__init__(f"Arguments {names} have different sizes in batch (first) dimension: {shapes}")
-
-
-class ShapeError(Exception):
-    """
-    Exception raised when an argument has less than two dimensions.
-
-    Parameters
-    ----------
-    arg_name: str
-        Argument name.
-    arg_shape: tuple of int
-        Argument shape.
-
-    """
-
-    def __init__(self, arg_name: str, arg_shape: tuple[int, ...]):
-        super().__init__(
-            f"Argument {arg_name} must have at least two dimensions but has shape {arg_shape}",
-        )
-
-
-class BaseModel(ABC):
-    """
-    Abstract base class for models.
-
-    Cannot be instantiated on its own.
-    Can only be used as a parent class to create custom model classes.
-    Subclasses must override the abstract :attr:`parameter_names` property.
-
-    Attributes
-    ----------
-    parameter_names
-
-    Examples
-    --------
-    Create a custom model class that inherits from the base class:
-
-    >>> class CustomModel(BaseModel):
-    ...     @property
-    ...     def parameter_names(self):
-    ...         return ["a", "b"]
-    >>> model = CustomModel()
-    >>> print(model.parameter_names)
-    ['a', 'b']
-
-    """
-
-    @property
-    @abstractmethod
-    def parameter_names(self) -> list[str]:
-        """A list with names of parameters that are used by the model."""
-
-
-class BaseResponse(BaseModel, Generic[S]):
+class BaseResponse(ModelProtocol, Generic[S]):
     """
     Generic abstract base class for response models.
 
@@ -118,7 +44,7 @@ class BaseResponse(BaseModel, Generic[S]):
         """
 
 
-class BaseEncoder(BaseModel, Generic[S]):
+class BaseEncoder(ModelProtocol, Generic[S]):
     """
     Generic abstract base class for encoding model responses.
 
@@ -158,111 +84,7 @@ class BaseEncoder(BaseModel, Generic[S]):
         """
 
 
-class BaseImpulse(BaseModel):
-    """
-    Abstract base class for impulse response models.
-
-    Cannot be instantiated on its own.
-    Can only be used as a parent class to create custom impulse response models.
-    Subclasses must override the abstract :meth:`__call__` method.
-
-    Parameters
-    ----------
-    duration : float, default=32.0
-        The duration of the impulse response (in seconds).
-    offset : float, default=0.0001
-        The offset of the impulse response (in seconds). By default a very small offset is added to prevent infinite
-        response values at t = 0.
-    resolution : float, default=1.0
-        The time resultion of the impulse response (in seconds), that is the number of points per second at which the
-        impulse response function is evaluated.
-    norm : str, optional, default="sum"
-        The normalization of the response. Can be `"sum"` (default), `"mean"`, `"max"`, `"norm"`, or `None`.
-        If `None`, no normalization is performed.
-    default_parameters : dict of float, optional
-        Dictionary with scalar default parameter values. Keys must be valid parameter names.
-
-    """
-
-    def __init__(
-        self,
-        duration: float = 32.0,
-        offset: float = 0.0001,
-        resolution: float = 1.0,
-        norm: str | None = "sum",
-        default_parameters: dict[str, float] | None = None,
-    ):
-        super().__init__()
-
-        self.duration = duration
-        self.offset = offset
-        self.resolution = resolution
-
-        # Check if norm arg is valid
-        if norm is not None:
-            _get_norm_fun(norm)
-
-        self.norm = norm
-
-        if default_parameters is not None:
-            if any(key not in self.parameter_names for key in default_parameters):
-                msg = "Invalid default parameter name, please provide valid parameter default parameter names"
-                raise ValueError(msg)
-
-            if any(not isinstance(val, float) for val in default_parameters.values()):
-                msg = "Default parameters must be single float values"
-                raise ValueError(msg)
-
-        self.default_parameters = default_parameters
-
-        self._frames: Tensor | None = None
-
-    @property
-    def num_frames(self) -> int:
-        """The total number of time frames at which the impulse response function is evaluated."""
-        return int(self.duration / self.resolution)
-
-    @property
-    def frames(self) -> Tensor:
-        """
-        The time frames at which the impulse response function is evaluated.
-
-        Time frames are linearly interpolated between `offset` and `duration` and have shape `(1, num_frames)`.
-
-        """
-        if self._frames is None:
-            self._frames = ops.expand_dims(ops.linspace(self.offset, self.duration, self.num_frames), 0)
-
-        return self._frames
-
-    def _join_default_parameters(self, parameters: pd.DataFrame) -> pd.DataFrame:
-        if self.default_parameters is not None:
-            parameters = parameters.copy()
-
-            for key, val in self.default_parameters.items():
-                parameters[key] = val
-
-        return parameters
-
-    @doc
-    @abstractmethod
-    def __call__(self, parameters: pd.DataFrame, dtype: str | None = None) -> Tensor:
-        """
-        Compute the impulse response.
-
-        Parameters
-        ----------
-        %(parameters)s
-        %(dtype)s
-
-        Returns
-        -------
-        %(predicted_response_2d)s
-
-        """
-
-
-class BaseTemporal(BaseModel):
+class BaseTemporal(ModelProtocol):
     """
     Abstract base class for temporal models.
 
@@ -292,7 +114,7 @@ class BaseTemporal(BaseModel):
         """
 
 
-class BaseComposite(BaseModel, Generic[S]):
+class BaseComposite(ModelProtocol, Generic[S]):
     """
     Generic abstract base class for creating composite models.
 
@@ -305,21 +127,22 @@ class BaseComposite(BaseModel, Generic[S]):
     Parameters
     ----------
     **models
-        Submodels to be combined into the composite model. All submodel classes must inherit from :class:`BaseModel`.
+        Submodels to be combined into the composite model. All submodel classes must inherit from
+        :class:`~prfmodel.utils.ModelProtocol`.
 
     Raises
     ------
     TypeError
-        If submodel classes do not inherit from :class:`BaseModel`.
+        If submodel classes do not inherit from :class:`~prfmodel.utils.ModelProtocol`.
 
     """
 
-    def __init__(self, **models: BaseModel | None):
+    def __init__(self, **models: ModelProtocol | None):
         super().__init__()
 
         for model in models.values():
-            if model is not None and not issubclass(model.__class__, BaseModel):
-                msg = "Model instance must inherit from BaseModel"
+            if model is not None and not isinstance(model, ModelProtocol):
+                msg = "Model instance must implement the 'parameter_names' property"
                 raise TypeError(msg)
 
         self.models = models
