+++ src/prfmodel/models/div_norm.py
@@ -3,18 +3,18 @@
 from typing import cast
 import pandas as pd
 from keras import ops
+from prfmodel.impulse import DerivativeTwoGammaImpulse
+from prfmodel.impulse import convolve_prf_impulse_response
+from prfmodel.impulse.base import BaseImpulse
 from prfmodel.stimuli.prf import PRFStimulus
 from prfmodel.typing import Tensor
 from prfmodel.utils import get_dtype
 from .base import BaseComposite
 from .base import BaseEncoder
-from .base import BaseImpulse
 from .base import BaseResponse
 from .base import BaseTemporal
 from .encoding import PRFStimulusEncoder
 from .gaussian import Gaussian2DPRFResponse
-from .impulse import DerivativeTwoGammaImpulse
-from .impulse import convolve_prf_impulse_response
 from .temporal import DivNormAmplitude
