+++ src/prfmodel/models/gaussian.py
@@ -5,6 +5,10 @@ import numpy as np
 import pandas as pd
 from keras import ops
 from prfmodel._docstring import doc
+from prfmodel.exceptions import BatchDimensionError
+from prfmodel.exceptions import ShapeError
+from prfmodel.impulse import DerivativeTwoGammaImpulse
+from prfmodel.impulse.base import BaseImpulse
 from prfmodel.stimuli import CFStimulus
 from prfmodel.stimuli import GridDimensionsError
 from prfmodel.stimuli import PRFStimulus
@@ -13,16 +17,12 @@ from prfmodel.utils import _EXPECTED_NDIM
 from prfmodel.utils import convert_parameters_to_tensor
 from prfmodel.utils import get_dtype
 from .base import BaseEncoder
-from .base import BaseImpulse
 from .base import BaseResponse
 from .base import BaseTemporal
-from .base import BatchDimensionError
-from .base import ShapeError
 from .composite import SimpleCFModel
 from .composite import SimplePRFModel
 from .encoding import CFStimulusEncoder
 from .encoding import PRFStimulusEncoder
-from .impulse import DerivativeTwoGammaImpulse
 from .temporal import BaselineAmplitude
