+++ tests/models/test_dog.py
@@ -3,13 +3,13 @@
 import numpy as np
 import pandas as pd
 import pytest
-from prfmodel.models.base import BaseImpulse
+from prfmodel.impulse import DerivativeTwoGammaImpulse
+from prfmodel.impulse.base import BaseImpulse
 from prfmodel.models.base import BaseTemporal
 from prfmodel.models.composite import CenterSurroundPRFModel
 from prfmodel.models.dog import DoG2DPRFModel
 from prfmodel.models.dog import init_dog_from_gaussian
 from prfmodel.models.gaussian import Gaussian2DPRFResponse
-from prfmodel.models.impulse import DerivativeTwoGammaImpulse
 from prfmodel.models.temporal import DoGAmplitude
 from prfmodel.stimuli.prf import PRFStimulus
 from tests.conftest import PRFStimulusSetup
