+++ tests/models/test_gaussian.py
@@ -5,11 +5,12 @@ import pandas as pd
 import pytest
 from pytest_regressions.num_regression import NumericRegressionFixture
 from scipy import stats
+from prfmodel.exceptions import BatchDimensionError
+from prfmodel.exceptions import ShapeError
+from prfmodel.impulse import DerivativeTwoGammaImpulse
+from prfmodel.impulse.base import BaseImpulse
 from prfmodel.models.base import BaseEncoder
-from prfmodel.models.base import BaseImpulse
 from prfmodel.models.base import BaseTemporal
-from prfmodel.models.base import BatchDimensionError
-from prfmodel.models.base import ShapeError
 from prfmodel.models.encoding import CFStimulusEncoder
 from prfmodel.models.encoding import PRFStimulusEncoder
 from prfmodel.models.gaussian import Gaussian2DPRFModel
@@ -20,7 +21,6 @@ from prfmodel.models.gaussian import GridMuDimensionsError
 from prfmodel.models.gaussian import _check_gaussian_args
 from prfmodel.models.gaussian import _expand_gaussian_args
 from prfmodel.models.gaussian import predict_gaussian_response
-from prfmodel.models.impulse import DerivativeTwoGammaImpulse
 from prfmodel.models.temporal import BaselineAmplitude
 from prfmodel.stimuli.cf import CFStimulus
 from prfmodel.stimuli.prf import GridDimensionsError
